basic kmp infra + db setup

+++ shared/src/commonMain/kotlin/org/glebkrasnopolin/goodhabit/models/ActionLog.kt
@@ -3,6 +3,6 @@ package org.glebkrasnopolin.goodhabit.models
 data class ActionLog (
     val type: ActionType,
     val timestamp: Long,
-    val value: Double? = null,
+    val value: Int? = null,
     // val actionTrendMoralEvaluation: Boolean?
 )
